Add files via upload
Creating A Blog In Django has interactive UI design using which users can see what others are posting. It also has an admin panel through which all the blog posts and users can be managed.

<br>
<br>

![1](PREVIEW/IMG1.png)
![2](PREVIEW/IMG2.png)
![3](PREVIEW/IMG3.png)
![4](PREVIEW/IMG4.png)
![5](PREVIEW/IMG5.png)
<br>


## Features 

- **Manage Blog** :– In this feature includes the CRUD operation in a blog or content you create like adding, editing and deleting content of the blog
- **Login System** :- In this feature the admin can login to the system and manage all the feature of the system.
- **Blog** :- In this method which is the main method of the system.
- **Media** :- In this method which you can found all the media that you are upload in the system.
- **Template** :- In this method which is the design of the system that consist of HTML,CSS and JavaScript.



 ## 🛠 Built With

* HTML
* CSS
* JAVASCRIPT
* PYTHON
* DJANGO
* DATABASE 

The system is built fully in Django Framework in back-end and HTML, CSS in front-end. It has full-featured user interface with all the functionalities


## ⭐️ Show your support 

Give a ⭐️ if you like this project!